pyproject.toml
[tool.poetry]
name = "foodie"
version = "0.1.0"
description = ""
authors = ["Shaibu Shaibu <s.shaibu.jnr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
uvicorn = "^0.14.0"
SQLAlchemy = "^1.4.17"
SQLAlchemy-Utils = "^0.37.6"
psycopg2 = "^2.8.6"
PyJWT = "^2.1.0"
inflection = "^0.5.1"
bcrypt = "^3.2.0"
passlib = "^1.7.4"
alembic = "^1.6.5"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.8.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
